rename to componentUtils and fix cityLines to draw between boundary cities
