+++ src/components/CityLines.js
@@ -4,7 +4,7 @@ const CityLines = () => {
   const boundaryCities = ['Sydney','Manila','Tokyo','SanFrancisco','LosAngeles'];
   let lines = [];
   drawNeighborLines(lines, boundaryCities);
-  drawBoundaryLines(lines, cities, boundaryCities);
+  lines = drawBoundaryLines(lines, cities, boundaryCities);
   return lines;
 };
