+++ src/utils/index.js
@@ -1,3 +1,3 @@
 export * from './cards';
 export * from './gameLogic';
-export * from './components';
+export * from './componentUtils';
